Add tests for xattr functions
Also move code to string_util to allow testing.  References #2810.
